% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bucharest}
\alias{bucharest}
\title{CRiSp example data for Bucharest}
\format{
A list with the following components:
\describe{
\item{boundary}{A sf object representing the administrative boundary of Bucharest.}
}
}
\source{
OpenStreetMap
}
\usage{
bucharest
}
\description{
Data extracted from OpenStreetMap for examples used in the CRiSp package.
}
\keyword{datasets}
