% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bucharest_delineation}
\alias{bucharest_delineation}
\title{CRiSp example delineation data for Bucharest}
\format{
A list of sf objects representing:
\describe{
\item{corridor}{The corridor boundaries of river Dâmbovița.}
\item{segments}{Corridor segments of river Dâmbovița.}
\item{riverspace}{River space of Dâmbovița.}
}
}
\usage{
bucharest_delineation
}
\description{
Delineation generated with \code{\link{bucharest_osm}} and \code{\link{bucharet_dem}} as input
}
\keyword{datasets}
