% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bucharest_dem}
\alias{bucharest_dem}
\title{CRiSp example DEM data for Bucharest}
\format{
TODO
}
\source{
Copernicus DEM 30
}
\usage{
bucharest_dem
}
\description{
Digital Elevation Model (DEM) for examples used in the CRiSp package.
}
\keyword{datasets}
