% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bucharest_osm}
\alias{bucharest_osm}
\title{CRiSp example OSM data for Bucharest}
\format{
A list with the following components:
\describe{
\item{boundary}{A sf object representing the administrative boundary of
Bucharest.}
}
}
\source{
OpenStreetMap
}
\usage{
bucharest_osm
}
\description{
Data extracted from OpenStreetMap for examples used in the CRiSp package.
}
\keyword{datasets}
