% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buffer_bbox}
\alias{buffer_bbox}
\title{Apply a buffer region to a bounding box}
\usage{
buffer_bbox(bbox, buffer)
}
\arguments{
\item{bbox}{Bonding box as a simple feature object}

\item{buffer}{Buffer region in meters}
}
\value{
Expanded bounding box as a simple feature object
}
\description{
If the input bbox is in lat/lon coordinates, the buffer is approximately
applied by first transforming the bbox in a suitable projected coordinate
reference system, expanding it with the given buffer, transforming it back to
the lat/lon system, and finally taking the bounding box of the obtained area.
}
