% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buffer_obj}
\alias{buffer_obj}
\title{Apply a buffer region to a sf object}
\usage{
buffer_obj(obj, buffer)
}
\arguments{
\item{buffer}{Buffer region in meters}

\item{a}{sf object}
}
\value{
Expanded sf object
}
\description{
If the input object is in lat/lon coordinates, the buffer is approximately
applied by first transforming the object in a suitable projected coordinate
reference system, expanding it with the given buffer.
}
