% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{build_river_network}
\alias{build_river_network}
\title{Build a spatial network from river centerlines}
\usage{
build_river_network(river, bbox = NULL)
}
\arguments{
\item{river}{A (MULTI)LINESTRING simple feature geometry representing the
river centerline}

\item{bbox}{Bounding box of the area of interest}
}
\value{
A \code{\link[sfnetworks:sfnetwork]{sfnetworks::sfnetwork}} object
}
\description{
If a bounding box is provided, only the river segments that intersect it are
considered. If the river intersects the bounding box multiple times, only the
longest intersecting segment will be considered.
}
\keyword{internal}
