% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_directory}
\alias{cache_directory}
\title{Return the cache directory used by the package}
\usage{
cache_directory()
}
\value{
The cache directory used by CRiSp.
}
\description{
By default, the cache directory is equal to \verb{$\{HOME\}/.cache/CRiSp}. A
different directory can be used by setting the environment variable
\code{CRISP_CACHE_DIRECTORY}. This can also be done by adding the following
line to the \code{.Renviron} file:
\verb{CRISP_CACHE_DIRECTORY=/path/to/crisp/cache/dir}.
}
