% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_weights}
\alias{calc_weights}
\title{Calculate the weights of the edges of a network based on their length}
\usage{
calc_weights(net)
}
\arguments{
\item{net}{A network object}
}
\value{
A network object with a new column 'weight' containing the length of
the edges
}
\description{
Calculate the weights of the edges of a network based on their length
}
