% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{check_cache}
\alias{check_cache}
\title{Check cache}
\usage{
check_cache()
}
\value{
\code{NULL}
}
\description{
A warning is raised if the cache size is > 100 MB or if it includes files
older than 30 days.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
check_cache()
\dontshow{\}) # examplesIf}
}
