% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_invalid_geometry}
\alias{check_invalid_geometry}
\title{Check and fix invalid geometries}
\usage{
check_invalid_geometry(sf_obj)
}
\arguments{
\item{sf_obj}{sf object}
}
\value{
sf object with valid geometries
}
\description{
Check and fix invalid geometries
}
\keyword{internal}
