% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Remove cache files}
\usage{
clear_cache(before_date = NULL)
}
\arguments{
\item{before_date}{Date before which cache files should be removed provided
as \code{Date}}
}
\value{
List of file paths of removed files
}
\description{
Remove files from cache directory either before a given date or entirely.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Clear all cache
clear_cache()

# Clear cache before given date
before_date <- as.Date("1-1-1999", "\%m-\%d-\%Y")
clear_cache(before_date)
\dontshow{\}) # examplesIf}
}
