% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{clip_and_filter}
\alias{clip_and_filter}
\title{Clip lines to the extent of the corridor, and select valid segment edges}
\usage{
clip_and_filter(lines, corridor, river)
}
\arguments{
\item{lines}{Candidate segment edges as a simple feature geometry}

\item{corridor}{The river corridor as a simple feature geometry}

\item{river}{The river centerline as a simple feature geometry}
}
\value{
Candidate segment edges as object of class \code{\link[sf:sfc]{sf::sfc_LINESTRING}}
}
\description{
Lines that intersect the river and that cross the corridor from side to side
are considered valid segment edges. We group valid segment edges that cross
the river in nearby locations, and select the shortest line per cluster.
From these candidate segment edges, we select the ultimate set of
non-intersecting lines by dropping the longest segments with most
intersections.
}
\keyword{internal}
