% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{cluster_crossings}
\alias{cluster_crossings}
\title{Group the river crossingns into clusters.}
\usage{
cluster_crossings(crossings, river, eps = 100)
}
\arguments{
\item{crossings}{Crossing edge geometries as a simple feature object}

\item{river}{The river geometry as a simple feature object}

\item{eps}{DBSCAN parameter referring to the size (radius) distance of the
neighborhood. Should approximate the distance between edges that we want
to consider as a single river crossing}
}
\value{
A simple feature geometry where the \code{cluster} column labels crossings
that are part of the same group
}
\description{
Create groups of edges that are crossing the river in nearby locations,
using a density-based clustering method (DBSCAN). This is to make sure that
edges representing e.g. different lanes of the same street are treated as
part of the same crossing.
}
