% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_river_features}
\alias{combine_river_features}
\title{Combine river centerline and surface}
\usage{
combine_river_features(river_centerline, river_surface)
}
\arguments{
\item{river_centerline}{River line as sfc_LINESTRING or sfc_MULTILINESTRING}

\item{river_surface}{River surface as sfc_POLYGON or sfc_MULTIPOLYGON}
}
\value{
Combined river as sfc_LINESTRING
}
\description{
Combine river centerline and surface
}
\examples{
river_centerline <- sf::st_sfc(
  sf::st_linestring(matrix(c(0, 0, 2, 2), ncol = 2, byrow = TRUE))
)
river_surface <- sf::st_sfc(
  sf::st_polygon(list(matrix(c(0, 0, 1, 0, 1, 1, 0, 1, 0, 0),
                             ncol = 2, byrow = TRUE)))
)
combine_river_features(river_centerline, river_surface)
}
