% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{corridor}
\alias{corridor}
\title{Delineate a river corridor on a spatial network.}
\usage{
corridor(
  network,
  river_centerline,
  river_surface,
  bbox,
  initial_method = "buffer",
  capping_method = "direct"
)
}
\arguments{
\item{network}{The spatial network to be used for the delineation}

\item{river_centerline}{A simple feature geometry representing the river
centerline}

\item{river_surface}{A simple feature geometry representing the river surface}

\item{bbox}{Bounding box defining the extent of the area of interest}

\item{initial_method}{The method employed to define the initial river
corridor geometry. See \code{\link[=initial_corridor]{initial_corridor()}} for the available methods}

\item{capping_method}{The method employed to connect the corridor edge end
points (i.e. to "cap" the corridor). See \code{\link[=cap_corridor]{cap_corridor()}} for
the available methods
TODO: add input arguments for the initialization method}
}
\value{
A simple feature geometry representing the river corridor
}
\description{
Delineate a river corridor on a spatial network.
}
