% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{corridor_end_points}
\alias{corridor_end_points}
\title{Find the corridor end points.}
\usage{
corridor_end_points(river, bbox)
}
\arguments{
\item{river}{A simple feature geometry representing the river centerline}

\item{bbox}{Bounding box defining the extent of the area of interest}
}
\value{
A simple feature geometry including a pair of points
}
\description{
Using the river center line and the bounding box of the area of interest,
determine the extremes (start and end point) of the river corridor.
}
