% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{create_network}
\alias{create_network}
\title{Create a network from a line strings}
\usage{
create_network(data, crs = NULL)
}
\arguments{
\item{data}{A data frame with a column named 'highway' containing
line strings}

\item{crs}{A coordinate reference system as an epsg code, e.g. 4326 for WGS84}
}
\value{
A network object
}
\description{
Create a network from a line strings
}
