% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\docType{data}
\name{default_stac_dem}
\alias{default_stac_dem}
\title{Endpoint and collection ID of the default STAC colleciton where to access
digital elevation model (DEM) data. This is the global Copernicus DEM 30
dataset hosted on AWS, as listed in the EarthSearch STAC API endpoint.
Note that AWS credentials need to be setup in order to access the data (not
the catalog).
References:
\itemize{
\item \href{https://element84.com/earth-search/}{EarthSearch STAC API}
\item \href{https://spacedata.copernicus.eu/collections/copernicus-digital-elevation-model}{Copernicus DEM}
\item \href{https://copernicus-dem-30m.s3.amazonaws.com/readme.html}{AWS Copernicus DEM datasets}
\item \href{https://docs.sentinel-hub.com/api/latest/static/files/data/dem/resources/license/License-COPDEM-30.pdf}{Data license}
}}
\format{
An object of class \code{list} of length 2.
}
\usage{
default_stac_dem
}
\description{
Endpoint and collection ID of the default STAC colleciton where to access
digital elevation model (DEM) data. This is the global Copernicus DEM 30
dataset hosted on AWS, as listed in the EarthSearch STAC API endpoint.
Note that AWS credentials need to be setup in order to access the data (not
the catalog).
References:
\itemize{
\item \href{https://element84.com/earth-search/}{EarthSearch STAC API}
\item \href{https://spacedata.copernicus.eu/collections/copernicus-digital-elevation-model}{Copernicus DEM}
\item \href{https://copernicus-dem-30m.s3.amazonaws.com/readme.html}{AWS Copernicus DEM datasets}
\item \href{https://docs.sentinel-hub.com/api/latest/static/files/data/dem/resources/license/License-COPDEM-30.pdf}{Data license}
}
}
\keyword{datasets}
