% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate.R
\name{delineate_city_river}
\alias{delineate_city_river}
\title{Delineate a corridor around a river}
\usage{
delineate_city_river(
  city_name,
  river_name,
  corridor = TRUE,
  segments = TRUE,
  riverspace = TRUE
)
}
\arguments{
\item{city_name}{A character vector of length one.}

\item{river_name}{A character vector of length one.}

\item{corridor}{Whether to carry out the corridor delineation. Default is
TRUE.}

\item{segments}{Whether to carry out the corridor segmentation.
Default is TRUE.}

\item{riverspace}{Whether to carry out the riverspace delineation.
Default is TRUE.}
}
\value{
A list with the valley, corridor, segments, and riverspace
geometries as \code{\link[sf:sfc]{sf::sfc_POLYGON}} objects.
}
\description{
This is a convenience function used for quick delineation. With only the city
name and river name as input, it uses default delineation parameters, it
retrieves OSM and DEM data and returns a list with all three delineations.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
delineate_city_river("Bucharest", "Dâmbovița"))
\dontshow{\}) # examplesIf}
}
