% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate.R
\name{delineate_segments}
\alias{delineate_segments}
\title{Delineate segments of a river corridor.}
\usage{
delineate_segments()
}
\value{
A simple feature geometry
}
\description{
Delineate segments of a river corridor.
}
