% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{delineate_valley}
\alias{delineate_valley}
\title{Extract the river valley from the DEM}
\usage{
delineate_valley(dem, river)
}
\arguments{
\item{dem}{Digital elevation model of the region}

\item{river}{A simple feature geometry representing the river}
}
\value{
River valley as a simple feature geometry
}
\description{
The slope of the digital elevation model (DEM) is used as friction (cost)
surface to compute the cost distance from any grid cell of the raster to
the river. A characteristic value (default: the mean) of the cost distance
distribution in a region surrounding the river (default: a buffer region of
2 km) is then calculated, and used to threshold the cost-distance surface.
The resulting area is then "polygonized" to obtain the valley boundary as a
simple feature geometry.
}
\examples{
delineate_valley(terra::unwrap(CRiSpData::bucharest_dem),
                 CRiSpData::bucharest_osm$river_centerline)
}
