% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{dem_to_cog}
\alias{dem_to_cog}
\title{Write DEM to cloud optimized GeoTIFF file as specified location}
\usage{
dem_to_cog(dem, fpath, output_directory = NULL)
}
\arguments{
\item{dem}{DEM of class \code{SpatRaster} to write to file}

\item{fpath}{filepath for output. If no output directory is specified
(see below) fpath is parsed to determine the output directory}

\item{output_directory}{where file should be written.
If specified fpath is treated as filename only.}
}
\value{
\code{SpatRaster} object of cloud optimised GeoTIFF
}
\description{
Write DEM to cloud optimized GeoTIFF file as specified location
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bucharest_dem <- get_dem_example_data()
dem_to_cog(bucharest_dem, "bucharest_dem.tif", output_directory = ".")
\dontshow{\}) # examplesIf}
}
