% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_line}
\alias{extend_line}
\title{Extend a line by a given distance}
\usage{
extend_line(line, distance)
}
\arguments{
\item{line}{Linestring object}

\item{dist}{Distance to extend the line}
}
\value{
Extended linestring object
}
\description{
Extend a line by a given distance
}
