% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{flatten_network}
\alias{flatten_network}
\title{Flatten a network by adding points at apparent intersections}
\usage{
flatten_network(network)
}
\arguments{
\item{network}{A network object}
}
\value{
A network object with additional points at intersections
}
\description{
This function adds nodes at intersections between network edges that
cross each other but do not have vertices at the intersection.
}
