% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{get_cd_char}
\alias{get_cd_char}
\title{Get characteristic value of distribution of cost distance}
\usage{
get_cd_char(cd, method = "mean")
}
\arguments{
\item{cd}{cost distance raster data}

\item{method}{function used to derive caracteristic value (mean)}
}
\value{
characteristic value of cd raster
}
\description{
Get characteristic value of distribution of cost distance
}
