% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{get_corridor}
\alias{get_corridor}
\title{Delineate a river corridor on a spatial network.}
\usage{
get_corridor(
  network,
  river_centerline,
  river_surface,
  bbox,
  initial_method = "buffer",
  buffer = 1000,
  dem = NULL,
  capping_method = "direct"
)
}
\arguments{
\item{network}{The spatial network to be used for the delineation}

\item{river_centerline}{A simple feature geometry representing the river
centerline}

\item{river_surface}{A simple feature geometry representing the river surface}

\item{bbox}{Bounding box defining the extent of the area of interest}

\item{initial_method}{The method employed to define the initial river
corridor geometry. See \code{\link[=initial_corridor]{initial_corridor()}} for the available methods}

\item{buffer}{Buffer region to add to the river geometry to setup the initial
corridor (only used if \code{initial_method} is \code{"buffer"})}

\item{dem}{Digital elevation model (DEM) of the region (only used if
\code{initial_method} is \code{"valley"})}

\item{capping_method}{The method employed to connect the corridor edge end
points (i.e. to "cap" the corridor). See \code{\link[=cap_corridor]{cap_corridor()}} for
the available methods}
}
\value{
A simple feature geometry representing the river corridor
}
\description{
Delineate a river corridor on a spatial network.
}
