% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{get_corridor_edges}
\alias{get_corridor_edges}
\title{Split corridor along the river to find edges on the two banks}
\usage{
get_corridor_edges(corridor, river_centerline)
}
\arguments{
\item{corridor}{The river corridor as a simple feature geometry}

\item{river_centerline}{The river centerline as a simple feature geometry}
}
\value{
Corridor edges as a simple feature geometry
}
\description{
Split corridor along the river to find edges on the two banks
}
\keyword{-}
\keyword{For}
\keyword{actually}
\keyword{complex}
\keyword{corridor}
\keyword{geometries,}
\keyword{here}
\keyword{internal}
\keyword{linestrings}
\keyword{longest}
\keyword{might}
\keyword{multiple}
\keyword{return}
\keyword{river}
\keyword{segments}
\keyword{select}
\keyword{splitting}
\keyword{the}
\keyword{two}
