% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{get_cost_distance}
\alias{get_cost_distance}
\title{Derive cost distance function from masked slope}
\usage{
get_cost_distance(slope, river, target = 0)
}
\arguments{
\item{slope}{raster of slope data}

\item{river}{vector data of river}

\item{target}{value for cost distance calculation}
}
\value{
raster of cost distance
}
\description{
Derive cost distance function from masked slope
}
\keyword{internal}
