% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{get_crossing_edges}
\alias{get_crossing_edges}
\title{Identify network edges that are crossing the river.}
\usage{
get_crossing_edges(network, river)
}
\arguments{
\item{network}{A spatial network object}

\item{river}{The river geometry as a simple feature object}
}
\value{
Indices of the edges crossing the river
}
\description{
Identify network edges that are crossing the river.
}
