% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{get_dem}
\alias{get_dem}
\title{Access digital elevation model (DEM) for a given region}
\usage{
get_dem(
  aoi,
  osm,
  dem_source = "STAC",
  stac_endpoint = NULL,
  stac_collection = NULL,
  force_download = FALSE
)
}
\arguments{
\item{aoi}{A list of delineation parameters, including \verb{$dem_buffer} used
to expand the area of interest covered by the network and \verb{$crs} for the
CRS which to transform the DEM to}

\item{osm}{A list with OpenStreetMap data sets for the a location, as
objects of class \code{\link[sf:sfc]{sf::sfc}}}

\item{dem_source}{Source of the DEM:
\itemize{
\item If "STAC" (default), DEM tiles are searched on a SpatioTemporal Asset
Catalog (STAC) end point, then accessed and mosaicked to the area of
interest
}}

\item{stac_endpoint}{URL of the STAC API endpoint (only used if \code{dem_source}
is \code{"STAC"}). For more info, see \code{\link[=get_stac_asset_urls]{get_stac_asset_urls()}}}

\item{stac_collection}{Identifier of the STAC collection to be queried (only
used if \code{dem_source} is \code{"STAC"}). For more info, see
\code{\link[=get_stac_asset_urls]{get_stac_asset_urls()}}}

\item{force_download}{Download data even if cached data is available}
}
\value{
DEM as a terra \code{SpatRaster} object
}
\description{
Access digital elevation model (DEM) for a given region
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Define delineation parameters and get OSM data within area of interest
aoi <- define_aoi("Bucharest", "Dâmbovița")
osm <- get_osmdata(aoi)

# Get DEM with default values
dem <- get_dem(aoi, osm)

# Get DEM from custom STAC endpoint
get_dem(aoi, osm,
        stac_endpoint = "some endpoint",
        stac_collection = "some collection")

# Specify CRS
get_dem(bb, crs = crs)
\dontshow{\}) # examplesIf}
}
