% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_dem_cache_filepath}
\alias{get_dem_cache_filepath}
\title{Get file path where to cache digital elevation model (DEM) data}
\usage{
get_dem_cache_filepath(tile_urls, bbox)
}
\arguments{
\item{tile_urls}{URL-paths where to reach the DEM tiles}

\item{bbox}{A bounding box}
}
\value{
A character string representing the file path
}
\description{
The function returns the file path where to serialize a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}
object representing the DEM as retrieved from a set of tiles reachable at the
given URLs, cropped and merged for the given bounding box. The directory used
is the one returned by \code{\link[=cache_directory]{cache_directory()}}.
}
