% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_example_cache_filepath}
\alias{get_example_cache_filepath}
\title{Get file path where to cache example data files}
\usage{
get_example_cache_filepath(filename)
}
\arguments{
\item{filename}{The name of the file to be retrieved from the data repository}
}
\value{
A character string representing the file path
}
\description{
The function returns the file path where to store example data files that
are retrieved from a data repository (see \code{\link[=get_osm_example_data]{get_osm_example_data()}} and
\code{\link[=get_dem_example_data]{get_dem_example_data()}}). The directory used is the one returned by
\code{\link[=cache_directory]{cache_directory()}}.
}
\keyword{internal}
