% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\name{get_example_data_file}
\alias{get_example_data_file}
\title{Retrieve an example data file from the data repository}
\usage{
get_example_data_file(filename, force_download = FALSE)
}
\arguments{
\item{force_download}{Download data even if cached data is available}
}
\value{
A character string representing the file path
}
\description{
Store the file in the cache directory, for subsequent reuse.
}
\keyword{internal}
