% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{get_intersecting_edges}
\alias{get_intersecting_edges}
\title{Identify network edges that are intersecting a geometry}
\usage{
get_intersecting_edges(network, geometry, index = FALSE)
}
\arguments{
\item{network}{A spatial network object of class \code{\link[sfnetworks:sfnetwork]{sfnetworks::sfnetwork}}}

\item{geometry}{An object of class \code{\link[sf:sfc]{sf::sfc}}}

\item{index}{Whether to return the indices of the matchin edges or the
geometries}
}
\value{
Indices or geometries of the edges intersecting the given geometry
of class \code{\link[sf:sfc]{sf::sfc_LINESTRING}}
}
\description{
Identify network edges that are intersecting a geometry
}
\keyword{internal}
