% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_latlon}
\alias{get_latlon}
\title{Get the latitude and longitude of an sf object}
\usage{
get_latlon(x)
}
\arguments{
\item{x}{Object of sf class sf, sfc or sfg}
}
\value{
An sf object with the latitude and longitude
}
\description{
Get the latitude and longitude of an sf object
}
