% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.R
\name{get_osm}
\alias{get_osm}
\title{Retrieve OpenStreetMap data for a given location}
\usage{
get_osm(
  aoi,
  city_boundary = TRUE,
  network = TRUE,
  buildings = TRUE,
  force_download = FALSE
)
}
\arguments{
\item{aoi}{A list of delineation parameters}

\item{city_boundary}{A logical indicating if the city boundary should be
retrieved. Default is TRUE.}

\item{network}{A logical indicating if the spatial network should be
retrieved. Default is TRUE.}

\item{buildings}{A logical indicating if buildings should be retrieved.
Default is TRUE.}

\item{force_download}{Download data even if cached data is available}
}
\value{
A list with the retrieved OpenStreetMap data sets for the
given location, as objects of class \code{\link[sf:sfc]{sf::sfc}}
}
\description{
Retrieve OpenStreetMap data for a given location, including
the city boundary, the river centerline and surface, the streets, the
railways, and the buildings
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Set parameters
city <- "Bucharest"
river <- "Dâmbovița"

# Define AOI parameters
aoi <- define_aoi(city, river)

# Get OSM data with defaults
get_osm(aoi)

# Get OSM data without city boundary
get_osm(aoi, city_boundary = FALSE)

# Use custom network buffer to get streets and railways
aoi2 <- aoi
aoi2$network_buffer = 3500
get_osm(aoi2)

# Use custom buffer to get buildings
aoi3 <- aoi
aoi3$buildings_buffer = 150
get_osm(aoi3)

# Use custom CRS
aoi4 <- aoi
aoi4$crs <- "EPSG:31600"
get_osm(aoi4)

# Avoid getting OSM data from cache
get_osm(city_name = city, river_name = river, force_download = TRUE)
\dontshow{\}) # examplesIf}
}
