% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata.R
\name{get_osm_bb}
\alias{get_osm_bb}
\title{Get the bounding box of a city}
\usage{
get_osm_bb(city_name)
}
\arguments{
\item{city_name}{The name of the city}
}
\value{
A bbox object with the bounding box of the city
}
\description{
Get the bounding box of a city
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_osm_bb("Bucharest")
\dontshow{\}) # examplesIf}
}
