% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retreive-osm.R
\name{get_osm_city_boundary}
\alias{get_osm_city_boundary}
\title{Get the city boundary from OpenStreetMap}
\usage{
get_osm_city_boundary(city_name)
}
\arguments{
\item{city_name}{A character string with the name of the city}
}
\value{
An sf object with the city boundary
}
\description{
Get the city boundary from OpenStreetMap
}
\examples{
get_osm_city_boundary("Bucharest")
}
