% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_osm_id_cache_filepath}
\alias{get_osm_id_cache_filepath}
\title{Get file path where to cache results of an Overpass API query by OSM id}
\usage{
get_osm_id_cache_filepath(type, id)
}
\arguments{
\item{type}{A character string with the OSM element type (e.g. "relation",
"way", "node")}

\item{id}{A character string with the OSM element id}
}
\value{
A character string representing the file path
}
\description{
The function returns the file path where to serialize an osmdata_sf object
for a given OSM element type and id. The directory used is the one returned
by \code{\link[=cache_directory]{cache_directory()}}.
}
\keyword{internal}
