% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_osmdata_cache_filepath}
\alias{get_osmdata_cache_filepath}
\title{Get the file path where to cache results of an Overpass API query}
\usage{
get_osmdata_cache_filepath(key, value, bbox)
}
\arguments{
\item{key}{A character string with the key to filter the data}

\item{value}{A character string with the value to filter the data}

\item{bbox}{A bounding box}
}
\value{
A character string representing the file path
}
\description{
The function returns the file path where to serialize an osdata_sf object
for a given key:value pair and a bounding box. The directory used is the one
returned by \code{\link[=cache_directory]{cache_directory()}}.
}
