% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata.R
\name{get_river_aoi}
\alias{get_river_aoi}
\title{Get an area of interest (aoi) around a river, cropping to the bounding box of
the city}
\usage{
get_river_aoi(river, city_bbox, buffer_distance = NULL)
}
\arguments{
\item{river}{A list with the river centreline and surface}

\item{buffer_distance}{Buffer size around the river}

\item{bbox}{Bounding box around the city}
}
\value{
An sf object or bbox
}
\description{
Get an area of interest (aoi) around a river, cropping to the bounding box of
the city
}
\examples{
bb <- get_osm_bb("Bucharest")
river <- get_osm_river(bb, "Dâmbovița")
get_river_aoi(river, bb, buffer_distance = 100)
}
