% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{get_river_banks}
\alias{get_river_banks}
\title{Draw the regions corresponding to the two river banks}
\usage{
get_river_banks(river, width)
}
\arguments{
\item{river}{River spatial features provided as a \code{\link[sfnetworks:sfnetwork]{sfnetworks::sfnetwork}}
or \code{\link[sf:sf]{sf::sf}}/\code{\link[sf:sfc]{sf::sfc}} object.}

\item{width}{(Approximate) width of the regions}
}
\value{
A \code{\link[sf:sfc]{sf::sfc}} object with two polygon features
}
\description{
These are constructed as single-sided buffers around the river geometry (see
\code{\link[=river_buffer]{river_buffer()}} for the implementation and refinement steps).
}
