% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverspace.R
\name{get_riverspace}
\alias{get_riverspace}
\title{Get riverspace surrounding a river}
\usage{
get_riverspace(occluders, river, density = 1/50, rayno = 41, raylen = 100)
}
\arguments{
\item{occluders}{Geometry of occluders}

\item{river}{River centreline or surface}

\item{rayno}{Number of rays}

\item{raylen}{Length of rays}
}
\value{
Polygon geometry with the riverspace
}
\description{
Get riverspace surrounding a river
}
\examples{
get_riverspace(bucharest_osm$buildings, bucharest_osm$river_surface)
}
