% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{get_segments}
\alias{get_segments}
\title{Split a river corridor into segments}
\usage{
get_segments(corridor, network, river_centerline, angle_threshold = 90)
}
\arguments{
\item{corridor}{The river corridor as a simple feature geometry}

\item{network}{The spatial network to be used for the segmentation}

\item{river_centerline}{The river centerline as a simple feature geometry}

\item{angle_threshold}{Only consider angles above this threshold (in degrees)
to form continuous strokes in the network}
}
\value{
Segment polygons as a simple feature geometry
}
\description{
Segments are defined as corridor subregions separated by river-crossing
transversal lines that form continuous strokes in the network.
}
