% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate-corridor.R
\name{get_target_points}
\alias{get_target_points}
\title{Get start and end points of corridor edge on the network.}
\usage{
get_target_points(vertices, area, threshold = 0.001)
}
\arguments{
\item{vertices}{A simple feature geometry set of two points}

\item{area}{An area of interest as a simple feature}

\item{threshold}{A numeric value}
}
\value{
A simple feature geometry set of two points
}
\description{
TODO add description
}
