% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_utm_zone}
\alias{get_utm_zone}
\title{Get the UTM zone of a spatial object}
\usage{
get_utm_zone(x)
}
\arguments{
\item{x}{Object in any class accepted by \code{\link[=as_bbox]{as_bbox()}}}
}
\value{
The EPSG code of the UTM zone
}
\description{
Get the UTM zone of a spatial object
}
\examples{
# Get EPSG code for UTM zone of Bucharest
bb <- c(xmin = 25.97, ymin = 44.33, xmax = 26.23, ymax = 44.54)
get_utm_zone(bb)
}
