% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_utm_zone_epsg}
\alias{get_utm_zone_epsg}
\title{Get UTM zone from longitude}
\usage{
get_utm_zone_epsg(x)
}
\arguments{
\item{x}{An sf object}
}
\value{
The UTM zone
}
\description{
Get UTM zone from longitude
}
