% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_utm_zone_epsg_bb}
\alias{get_utm_zone_epsg_bb}
\title{Get UTM zone from longitude}
\usage{
get_utm_zone_epsg_bb(bb)
}
\arguments{
\item{x}{A matrix with longitude and latitude bounds}
}
\value{
The UTM zone
}
\description{
Get UTM zone from longitude
}
