% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_utm_zone_epsg_bbox}
\alias{get_utm_zone_epsg_bbox}
\title{Get UTM zone from bbox object}
\usage{
get_utm_zone_epsg_bbox(bb)
}
\arguments{
\item{bb}{A bbox object}
}
\value{
The UTM zone
}
\description{
Get UTM zone from bbox object
}
