% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_utm_zone_epsg_sf}
\alias{get_utm_zone_epsg_sf}
\title{Get UTM zone from sf object}
\usage{
get_utm_zone_epsg_sf(x)
}
\arguments{
\item{x}{An sf object}
}
\value{
The UTM zone
}
\description{
Get UTM zone from sf object
}
