% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{get_valley}
\alias{get_valley}
\title{Create vector/polygon representation of valley from dem and river polygon
for a crs}
\usage{
get_valley(dem, river, crs)
}
\arguments{
\item{dem}{of region}

\item{river}{vector/polygon representation of river area}

\item{crs}{coordiante reference system to use}
}
\value{
(multi)polygon representation of valley
area as st_geometry without holes
}
\description{
Create vector/polygon representation of valley from dem and river polygon
for a crs
}
