% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{get_valley_mask}
\alias{get_valley_mask}
\title{Select valley pixels from cost distance based on threshold}
\usage{
get_valley_mask(cd, thresh)
}
\arguments{
\item{cd}{cost distance raster}

\item{thresh}{threshold cost distance value below which pixels are assuemd
to belong to the valley}
}
\description{
Select valley pixels from cost distance based on threshold
}
