% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{get_valley_polygon}
\alias{get_valley_polygon}
\title{Create vector/polygon representation of valley without holes from raster mask}
\usage{
get_valley_polygon(valley_mask)
}
\arguments{
\item{valley_mask}{raster mask of valley pixels}
}
\value{
(multi)polygon representation of valley area as a simple feature
geometry without holes
}
\description{
Create vector/polygon representation of valley without holes from raster mask
}
