% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{get_valley_polygon_no_hole}
\alias{get_valley_polygon_no_hole}
\title{Remove possible holes from valley geometry}
\usage{
get_valley_polygon_no_hole(valley_polygon)
}
\arguments{
\item{valley_polygon}{st_geometry of valley region}
}
\value{
(multi)polygon geometry of valley
}
\description{
Remove possible holes from valley geometry
}
\keyword{internal}
