% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{get_valley_polygon_raw}
\alias{get_valley_polygon_raw}
\title{Create vector/polygon representation of valley raster mask}
\usage{
get_valley_polygon_raw(valley_mask)
}
\arguments{
\item{valley_mask}{raster mask of valley pixels}
}
\value{
polygon representation of valley area as st_geometry
}
\description{
Create vector/polygon representation of valley raster mask
}
\keyword{internal}
