% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate-corridor.R
\name{get_vertices}
\alias{get_vertices}
\title{Determine the end vertices of the initial river corridor.}
\usage{
get_vertices(aoi, corridor_initial)
}
\arguments{
\item{aoi}{Area of interest as a simple feature}

\item{corridor_initial}{Initial river corridor as a simple feature}
}
\value{
A simple feature geometry set of two points
}
\description{
Determine the "vertices" of the initial river corridor as the intersections
of the initial river corridor with the AoI boundary.
}
