% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{initial_corridor}
\alias{initial_corridor}
\title{Draw the initial geometry of a river corridor.}
\usage{
initial_corridor(river, method = "buffer", ..., bbox = NULL)
}
\arguments{
\item{river}{A simple feature geometry representing the river}

\item{method}{The method employed to draw the initial river corridor:
\itemize{
\item "buffer" (default): add a fixed buffer region to the river geometry
(see \code{\link[=river_buffer]{river_buffer()}})
}}

\item{...}{Additional arguments required by the function that implements the
chosen method (see \code{method})}

\item{bbox}{Bounding box defining the extent of the area of interest}
}
\value{
A simple feature geometry
}
\description{
Draw the initial geometry of a river corridor.
}
