% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{initial_edges}
\alias{initial_edges}
\title{Identify the initial edges of the river corridor}
\usage{
initial_edges(corridor_initial, regions)
}
\arguments{
\item{corridor_initial}{A simple feature geometry representing the area of
the initial corridor}

\item{regions}{A simple feature geometry representing the sub-regions formed
by cutting the area of interest along the river}
}
\value{
A simple feature geometry representing the initial corridor edges
}
\description{
These are defined by splitting the initial corridor boundary into the
sub-regions that the river forms in the area of interest
}
