% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{load_raster}
\alias{load_raster}
\title{Retrieve STAC records (of a DEM) corresponding to a list of asset urls,
crop and merge with a specified bounding box to create a dem of the
specified region}
\usage{
load_raster(bb, raster_urlpaths)
}
\arguments{
\item{bb}{A bounding box, provided either as a matrix (rows for "x", "y",
columns for "min", "max") or as a vector ("xmin", "ymin", "xmax", "ymax")}

\item{raster_urlpaths}{a list of STAC records to be retrieved}
}
\value{
A a merged dem from retrieved assets cropped to the bounding box
}
\description{
Retrieve STAC records (of a DEM) corresponding to a list of asset urls,
crop and merge with a specified bounding box to create a dem of the
specified region
}
