% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{mask_cost_distance}
\alias{mask_cost_distance}
\title{Mask out river regions incl. a buffer in cost distance raster data}
\usage{
mask_cost_distance(cd, river, buffer = 2000)
}
\arguments{
\item{cd}{cost distance raster}

\item{river}{vector/polygon}

\item{buffer}{size of buffer around river polygon to additionally mask}
}
\value{
cd raster with river+BUFFER pixels masked
}
\description{
Mask out river regions incl. a buffer in cost distance raster data
}
\keyword{internal}
