% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{mask_slope}
\alias{mask_slope}
\title{Mask slope raster, setting the slope to zero for the pixels overlapping
the river area.}
\usage{
mask_slope(slope, river, lthresh = 0.001, target = 0)
}
\arguments{
\item{slope}{raster data of slope}

\item{river}{vector/polygon data of river}

\item{lthresh}{lower numerival threshold to consider slope non-zero}

\item{target}{value to set for pixels overlapping river area}
}
\value{
updated slope raster
}
\description{
Mask slope raster, setting the slope to zero for the pixels overlapping
the river area.
}
