% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{merge_blocks}
\alias{merge_blocks}
\title{Merge a set of blocks to adjacent ones}
\usage{
merge_blocks(blocks, to_merge, method = "longest-intersection")
}
\arguments{
\item{blocks}{Simple feature geometry representing all the blocks}

\item{to_merge}{Indices of the blocks to merge}

\item{method}{Strategy for merging, see \code{\link[=merge_block]{merge_block()}}}
}
\value{
Blocks merged to the specified ones as a simple feature geometry
}
\description{
Adjacent blocks are defined as the blocks that are neighbours to the blocks
that need to be merged, and that intersect them via a (Multi)LineString. We
consider the blocks to merge one by one, from the smallest to the largest,
merging them to the other blocks recursively.
}
