% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{merge_streets}
\alias{merge_streets}
\title{Merge streets from polygons and lines}
\usage{
merge_streets(highways)
}
\arguments{
\item{highways}{A list object containing streets from OpenStreetMap}
}
\value{
A data frame with a column named 'highway' containing line strings
}
\description{
Merge streets from polygons and lines
}
