% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{nearest_node}
\alias{nearest_node}
\title{Find the node in a network that is closest to a target geometry.}
\usage{
nearest_node(network, target)
}
\arguments{
\item{network}{A network object of class \code{\link[sfnetworks:sfnetwork]{sfnetworks::sfnetwork}}}

\item{target}{An object of class \code{\link[sf:sf]{sf::sf}} or \code{\link[sf:sfc]{sf::sfc}}}
}
\value{
A node in the network as an object of class \code{\link[sf:sfc]{sf::sfc_POINT}}
}
\description{
Find the node in a network that is closest to a target geometry.
}
\keyword{internal}
