% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_intersects}
\alias{not_intersects}
\title{Check if two sf objects do not intersect}
\usage{
not_intersects(x, y)
}
\arguments{
\item{x}{sf object}

\item{y}{sf object}
}
\value{
A logical vector
}
\description{
Check if two sf objects do not intersect
}
