% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retreive-osm.R
\name{osm_bb}
\alias{osm_bb}
\title{Retrieve bounding box from OpenStreetMap}
\usage{
osm_bb(name)
}
\arguments{
\item{name}{A character string with the name of the place to retrieve the bounding box}
}
\value{
A list with the bounding box
}
\description{
Retrieve bounding box from OpenStreetMap
}
\examples{
osm_bb("Bucharest")
}
