% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retreive-osm.R
\name{osmdata_as_sf}
\alias{osmdata_as_sf}
\title{Retrieve OpenStreetMap data as sf object}
\usage{
osmdata_as_sf(key, value, bb)
}
\arguments{
\item{key}{A character string with the key to filter the data}

\item{value}{A character string with the value to filter the data}

\item{bb}{A list with the bounding box}
}
\value{
An sf object with the retrieved OpenStreetMap data
}
\description{
Retrieve OpenStreetMap data as sf object
}
