% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.R
\name{osmdata_as_sf_by_id}
\alias{osmdata_as_sf_by_id}
\title{Retrieve OpenStreetMap data as sf object for a specific feature type and id}
\usage{
osmdata_as_sf_by_id(type, id, force_download = FALSE)
}
\arguments{
\item{type}{A character string with the OSM element type ("relation",
"way", or "node")}

\item{id}{A character or numeric vector of length 1 with the OSM element id}

\item{force_download}{Download data even if cached data is available}
}
\value{
An \code{\link[osmdata:osmdata]{osmdata::osmdata}} object with the retrieved OpenStreetMap data
}
\description{
Results are cached, so that new queries with the same type and id will be
loaded from disk.
}
\keyword{internal}
