% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.R
\name{osmdata_id_query}
\alias{osmdata_id_query}
\title{Query the Overpass API for a specific feature type and id}
\usage{
osmdata_id_query(type, id)
}
\arguments{
\item{type}{A character string with the OSM element type}

\item{id}{A character or numeric vector of length 1 with the OSM element id}
}
\value{
An \code{\link[osmdata:osmdata]{osmdata::osmdata}} object with the retrieved OpenStreetMap data
}
\description{
Query the Overpass API for a specific feature type and id
}
\keyword{internal}
