% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata.R
\name{osmdata_query}
\alias{osmdata_query}
\title{Query the Overpass API for a key:value pair within a bounding box}
\usage{
osmdata_query(key, value, bb)
}
\arguments{
\item{key}{A character string with the key to filter the data}

\item{value}{A character string with the value to filter the data. If
value = "" means that you get all features available in OSM for the
specified bounding box}

\item{bb}{A bounding box, in lat/lon coordinates}
}
\value{
An sf object with the retrieved OpenStreetMap data
}
\description{
Query the Overpass API for a key:value pair within a bounding box
}
\keyword{internal}
