% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate.R
\name{plot.delineation}
\alias{plot.delineation}
\title{Plot a delineation object}
\usage{
\method{plot}{delineation}(x, ..., legend = TRUE)
}
\arguments{
\item{x}{An object of class \link{delineation}, which is typically the output of
the \code{\link[=delineate]{delineate()}} function.}

\item{...}{Not used; included for compatibility with the generic.}

\item{legend}{logical. If TRUE (default), a legend is added to the plot.}
}
\value{
\code{NULL}, invisibly. This function is called for its side effect of
producing a plot.
}
\description{
This S3 method provides a way to quickly visualise how the layers of a
delineation object fit together using base R graphics. A delineation object
typically includes the base layers \code{streets}, \code{railways},
\code{river_centerline} and \code{river_surface}, and the delineations of the
\code{valley}, \code{corridor}, \code{segments}, and \code{riverspace}. Depending on the
delineation object, some of the delineation layers may not be present and
thus will not be plotted.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bd <- delineate_city_river("Bucharest", "Dâmbovița")
plot(bd)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=geom_delineation]{geom_delineation()}} for a \code{ggplot2}-based alternative.
}
