% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_distance}
\alias{preprocess_distance}
\title{Pre-process one-dimensional distance input}
\usage{
preprocess_distance(x, arg_name = deparse(substitute(x)))
}
\arguments{
\item{x}{A distance value which may be \code{numeric}, or a \code{\link[units:units]{units::units}}
object, or any other class fir which the underlying \code{storage.mode} is
numeric.}

\item{arg_name}{Name of the argument, used in messages/errors.}
}
\value{
A vector of class \code{numeric} in meters.
}
\description{
Ensure that distance input is plain numeric in meters, regardless of
whether it is provided as a \code{units} object (e.g., a distance in kilometers),
a plain numeric or another vector-like class with numeric \code{storage.mode}.
}
\keyword{internal}
