% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineation-methods.R
\name{print.summary.delineation}
\alias{print.summary.delineation}
\title{Print a summary.delineation object}
\usage{
\method{print}{summary.delineation}(x)
}
\arguments{
\item{x}{An object of class \code{summary.delineation}, as returned by
\code{\link[=summary.delineation]{summary.delineation()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a summary.delineation object
}
