% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{read_data_from_cache}
\alias{read_data_from_cache}
\title{Read data from the cache directory}
\usage{
read_data_from_cache(filepath, unwrap = FALSE, quiet = FALSE)
}
\arguments{
\item{filepath}{Path of the file to deserialize as a character string}

\item{unwrap}{Whether the deserialized object should be "unpacked" (as
required by \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} objects)}

\item{quiet}{Omit warning on cache file being loaded}
}
\value{
Object deserialized
}
\description{
For the directory used for caching see \code{\link[=cache_directory]{cache_directory()}}.
}
\keyword{internal}
