% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{river_buffer}
\alias{river_buffer}
\title{Draw a corridor as a fixed buffer region around a river.}
\usage{
river_buffer(river, buffer_distance, bbox = NULL)
}
\arguments{
\item{river}{A simple feature geometry representing the river}

\item{buffer_distance}{Size of the buffer (in the river's CRS units)}

\item{bbox}{Bounding box defining the extent of the area of interest}
}
\value{
A simple feature geometry
}
\description{
The river geometry may consist of multiple spatial features, these are merged
after applying the buffer.
}
\examples{
river <- sf::st_sfc(sf::st_linestring(matrix(c(0, 0, 1, 1), ncol = 2)))
sf::st_crs(river) <- sf::st_crs("EPSG:4326")
river_buffer(river, 100)
}
