% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{river_valley}
\alias{river_valley}
\title{Draw a corridor as an estimate of the river valley.}
\usage{
river_valley(river, dem, bbox = NULL)
}
\arguments{
\item{river}{A simple feature geometry representing the river}

\item{dem}{The DEM of the area as a terra SpatRaster object}

\item{bbox}{Bounding box defining the extent of the area of interest}
}
\value{
A simple feature geometry
}
\description{
The valley is drawn on a digital elevation model (DEM) of the area,
see \code{\link[=get_valley]{get_valley()}} for details on the implementation.
}
