% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{segments}
\alias{segments}
\title{Split a river corridor into segments}
\usage{
segments(corridor, network, river_centerline)
}
\arguments{
\item{corridor}{The river corridor as a simple feature geometry}

\item{network}{The spatial network to be used for the segmentation}

\item{river_centerline}{The river centerline as a simple feature geometry}
}
\value{
Segment polygons as a simple feature geometry
}
\description{
Split a river corridor into segments
}
