% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{select_nonintersecting_lines}
\alias{select_nonintersecting_lines}
\title{Select non-intersecting line segments}
\usage{
select_nonintersecting_lines(lines, corridor)
}
\arguments{
\item{lines}{Candidate edge segment as a simple feature geometry}

\item{corridor}{The river corridor as a simple feature geometry}
}
\value{
A set of lines that do not intersect within the corridor geometry,
as a simple feature geometry
}
\description{
Recursively drop intersecting lines, starting from the line that form most
intersections with other geometries. When multilple lines form the same
number of intersections with other geometries, the longest line is discarded
first. Note that lines are allowed to intersect on the corridor boundary.
}
\keyword{internal}
