% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_units_like}
\alias{set_units_like}
\title{Set the units of x as the units of y}
\usage{
set_units_like(x, y)
}
\arguments{
\item{x}{x (can be unitless)}

\item{y}{y (can be unitless)}
}
\value{
Object x with units of y
}
\description{
Set the units of x as the units of y
}
\keyword{internal}
