% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{smooth_dem}
\alias{smooth_dem}
\title{Spatially smooth dem by (window) filtering}
\usage{
smooth_dem(dem, method = "median", window = 5)
}
\arguments{
\item{dem}{raster data of dem}

\item{method}{smoothing function to be used, e.g. "median",
as accepted by \code{\link[terra:focal]{terra::focal()}}}

\item{window}{size of smoothing kernel}
}
\value{
filtered dem
}
\description{
Spatially smooth dem by (window) filtering
}
\examples{
smooth_dem(terra::unwrap(bucharest_dem))
}
