% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{split}
\alias{split}
\title{Split a geometry along a (multi)linestring.}
\usage{
split(geometry, line)
}
\arguments{
\item{geometry}{Geometry to split}

\item{line}{Dividing (multi)linestring}
}
\value{
A simple feature object
}
\description{
Split a geometry along a (multi)linestring.
}
