% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{split_aoi}
\alias{split_aoi}
\title{Split the area of interest (AoI) by a river.}
\usage{
split_aoi(bbox, river)
}
\arguments{
\item{bbox}{Bounding box defining the extent of the area of interest}

\item{river}{A simple feature geometry representing the river centerline}
}
\value{
A simple feature geometry set of two areas of interest
}
\description{
Return the fragments produced. If more than two fragments are obtained, only
return the two largest fragments.
}
