% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{split_by}
\alias{split_by}
\title{Split a geometry along a (multi)linestring.}
\usage{
split_by(geometry, line, boundary = FALSE)
}
\arguments{
\item{geometry}{Geometry to split}

\item{line}{Dividing (multi)linestring}

\item{boundary}{Whether to return the split boundary instead of the regions}
}
\value{
An object of class \code{\link[sf:sfc]{sf::sfc}}
}
\description{
Split a geometry along a (multi)linestring.
}
\keyword{internal}
