% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{strokes}
\alias{strokes}
\title{Identify naturally continuos lines in a spatial network.}
\usage{
strokes(network, from_edge, angle_threshold = 0, flow_mode = FALSE)
}
\arguments{
\item{network}{A spatial network object}

\item{angle_threshold}{Consecutive line segments are considered part of the
same continuous line if the internal angle they form is larger than
`angle_threshold`` (in degrees). Should be in the range [0, 180) degrees}

\item{flow_mode}{If TRUE, line segments that belong to the same edge are not
split across strokes (even if they form internal angles smaller than
\code{angle_threshold}).}

\item{from_edges}{The indices of the network edges to be continued}
}
\value{
Stroke geometries as a simple feature object
}
\description{
Extend the specified network edges along the spatial network to form longer
strokes. See \code{\link[rcoins:stroke]{rcoins::stroke()}} for more details.
}
