% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate-corridor.R
\name{trim_network}
\alias{trim_network}
\title{Trim a network to an area of interest (AoI) and a river corridor.}
\usage{
trim_network(net, area, river_corridor)
}
\arguments{
\item{net}{A network object}

\item{area}{An area of interest as a simple feature}

\item{river_corridor}{A river corridor as a simple feature}
}
\value{
A network object
}
\description{
Trim a network to an area of interest (AoI) and a river corridor.
}
