% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{write_data_to_cache}
\alias{write_data_to_cache}
\title{Write data to the cache directory}
\usage{
write_data_to_cache(x, filepath, wrap = FALSE, quiet = FALSE)
}
\arguments{
\item{x}{Object to serialize to a file}

\item{filepath}{Path where to serialize x, as a character string}

\item{quiet}{Omit warning on cache file being written}

\item{unwrap}{Whether the object should be "packed" before serialization (as
required by \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} objects)}
}
\value{
\code{NULL} invisibly
}
\description{
For the directory used for caching see \code{\link[=cache_directory]{cache_directory()}}.
}
