% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bucharest_dambovita}
\alias{bucharest_dambovita}
\title{rcrisp example delineation data for Bucharest}
\format{
A list of sf objects representing:
\describe{
\item{valley}{The valley boundaries of river Dâmbovița.}
\item{corridor}{The corridor boundaries of river Dâmbovița.}
\item{segments}{Corridor segments of river Dâmbovița.}
\item{riverspace}{River space of Dâmbovița.}
}
}
\usage{
bucharest_dambovita
}
\description{
Delineation generated with rcrisp example data found at
\url{https://data.4tu.nl/datasets/f5d5e118-b5bd-4dfb-987f-fe10d1b9b386}
}
\keyword{datasets}
