% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bucharest_dem}
\alias{bucharest_dem}
\title{CRiSp example DEM data for Bucharest}
\format{
A PackedSpatRaster object. Run \code{\link[terra:wrap]{terra::unwrap()}} to extract the
DEM as a SpatRaster object
}
\source{
\url{https://dataspace.copernicus.eu/explore-data/data-collections/copernicus-contributing-missions/collections-description/COP-DEM}
}
\usage{
bucharest_dem
}
\description{
Copernicus GLO-30 Digital Elevation Model (DEM) cropped and retiled to cover
the city of Bucharest. Used for examples and vignettes in the CRiSp package.
}
\keyword{datasets}
