% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bucharest_osm}
\alias{bucharest_osm}
\title{CRiSp example OSM data for Bucharest}
\format{
A list of sf objects representing:
\describe{
\item{bb}{The city bounding box.}
\item{boundary}{The administrative boundary of Bucharest.}
\item{river_centerline}{The Dâmbovița river centerline.}
\item{river_surface}{The Dâmbovița river area.}
\item{aoi_network}{The area of interest (AoI) to include the network around
the river.}
\item{streets}{The street network.}
\item{railways}{The railway network.}
\item{aoi_buildings}{The area of interest (AoI) to include buildings around
the river.}
\item{buildings}{The buildings in the corridor.}
}
}
\source{
\url{https://www.openstreetmap.org/about}
}
\usage{
bucharest_osm
}
\description{
Data extracted from OpenStreetMap for examples used in the CRiSp package.
All datasets are provided in a projected coordinate reference system
(UTM 35), with exception for the bounding box, which is provided as latitude/
longitude coordinates (WGS84).
}
\keyword{datasets}
