% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buffer}
\alias{buffer}
\title{Apply a buffer region to a sf object}
\usage{
buffer(obj, buffer_distance, ...)
}
\arguments{
\item{obj}{A sf object}

\item{buffer_distance}{Buffer distance in meters}

\item{...}{Optional parameters passed on to \code{\link[sf:geos_unary]{sf::st_buffer()}}}
}
\value{
Expanded sf object
}
\description{
If the input object is in lat/lon coordinates, the buffer is approximately
applied by first transforming the object in a suitable projected CRS,
expanding it with the given buffer, and then transforming
it back to the lat/lon system.
}
