% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{build_river_network}
\alias{build_river_network}
\title{Build a spatial network from river centerlines}
\usage{
build_river_network(river, aoi = NULL)
}
\arguments{
\item{river}{A simple feature geometry representing the river centerline}

\item{aoi}{Area of interest, provided as a bounding box or as a polygon.}
}
\value{
A \code{\link[sfnetworks:sfnetwork]{sfnetworks::sfnetwork}} object
}
\description{
If an area of interest (aoi) is provided, only the river segments that
intersects it are considered. If the river intersects the area of interest
multiple times, only the longest intersecting segment will be considered.
}
\keyword{internal}
