% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate-corridor.R
\name{cap_corridor}
\alias{cap_corridor}
\title{Cap corridor edges with a city boundary.}
\usage{
cap_corridor(corridor_edges, river, crs, bb, cap = "city")
}
\arguments{
\item{corridor_edges}{Edge of the corridor as a simple feature}

\item{river}{River centerline as a simple feature}

\item{crs}{A coordinate reference system as an epsg code, e.g. 4326 for WGS84}

\item{bb}{A bounding box as a matrix with 4 elements: xmin, ymin, xmax, ymax}

\item{cap}{Character string with the type of cap to be used.
Default is "city"}
}
\value{
A simple feature geometry
}
\description{
Cap corridor edges with a city boundary.
}
