% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_invalid_geometry}
\alias{check_invalid_geometry}
\title{Check and fix invalid geometries}
\usage{
check_invalid_geometry(sf_obj)
}
\arguments{
\item{sf_obj}{An object of class \code{\link[sf:sf]{sf::sf}} or \code{\link[sf:sfc]{sf::sfc}}}
}
\value{
\code{\link[sf:sf]{sf::sf}} or \code{\link[sf:sfc]{sf::sfc}} object with valid geometries
}
\description{
Check and fix invalid geometries
}
\keyword{internal}
