% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate-corridor.R
\name{clean_network}
\alias{clean_network}
\title{Clean a street network by subdividing edges and removing pseudo-nodes.}
\usage{
clean_network(net)
}
\arguments{
\item{net}{A network object}
}
\value{
A cleaned network object
}
\description{
Subdivide edges by \href{https://luukvdmeer.github.io/sfnetworks/articles/sfn02_preprocess_clean.html#subdivide-edges}{adding missing nodes},
remove \href{https://luukvdmeer.github.io/sfnetworks/articles/sfn02_preprocess_clean.html#smooth-pseudo-nodes}{pseudo-nodes},
and keep only the main connected component of the network.
}
