% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{clip_and_filter}
\alias{clip_and_filter}
\title{Clip lines to the extent of the corridor, and filter valid segment edges}
\usage{
clip_and_filter(lines, corridor, river_centerline)
}
\arguments{
\item{lines}{Candidate segment edges as a simple feature geometry}

\item{corridor}{The river corridor as a simple feature geometry}

\item{river_centerline}{The river centerline as a simple feature geometry}
}
\value{
Candidate segment edges as a simple feature geometry
}
\description{
Lines that intersect the river only once and that cross the corridor from
side to side are considered valid segment edges. We group valid segment edges
that cross the river in nearby locations, and select the shortest line per
cluster.
}
