% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{corridor_edge}
\alias{corridor_edge}
\title{Draw a corridor edge on the spatial network.}
\usage{
corridor_edge(network, end_points, target_edge, exclude_area = NULL)
}
\arguments{
\item{network}{The spatial network used for the delineation}

\item{end_points}{Target start- and end-point}

\item{target_edge}{Target edge geometry to follow in the delineation}

\item{exclude_area}{Region that we aim to exclude from the delineation}
}
\value{
A simple feature geometry representing the edge (i.e. a linestring)
}
\description{
Draw a corridor edge on the spatial network.
}
