% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{corridor_end_points}
\alias{corridor_end_points}
\title{Find the corridor end points.}
\usage{
corridor_end_points(river_network, spatial_network, regions)
}
\arguments{
\item{river_network}{A \code{\link[sfnetworks:sfnetwork]{sfnetworks::sfnetwork}} object representing the
river centerline}

\item{spatial_network}{A \code{\link[sfnetworks:sfnetwork]{sfnetworks::sfnetwork}} object representing the
spatial network used for the delineation}

\item{regions}{A simple feature geometry representing the two river sides}
}
\value{
An \code{\link[sf:sfc]{sf::sfc_POINT}} geometry including a pair of points
}
\description{
Determine the extremes (end points) of the river corridor using the network
built from the river center line features (see \code{\link[=build_river_network]{build_river_network()}} and
the spatial network used for the delineation. The end points are selected as
the two furthest river crossings of the spatial network that connect the
sub-networks for each river sides.
}
\keyword{internal}
