% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{define_aoi}
\alias{define_aoi}
\title{Define delineation parameters within an area of interest}
\usage{
define_aoi(
  city_name,
  river_name,
  crs = NULL,
  network_buffer = 3000,
  dem_buffer = 2500,
  buildings_buffer = 100
)
}
\arguments{
\item{city_name}{A character vector of length one}

\item{river_name}{A character vector of length one}

\item{crs}{The projected Coordinate Reference System (CRS) to use. If not
provided, the suitable Universal Transverse Mercator (UTM) CRS is selected}

\item{network_buffer}{Add a buffer (an integer in meters) around
river to retrieve additional data (streets, railways, etc.).
Default is 3000 m.}

\item{dem_buffer}{Size of the buffer region (in meters) around the spatial
network to retrieve the DEM.}

\item{buildings_buffer}{Add a buffer (an integer in meters) around the
river to retrieve additional data (buildings). Default is 100 m.}
}
\value{
A list with delineation parameter values, namely \code{city_name} and
\code{river_name} as character vectors of length one, \code{bb} as object of class
\code{\link[sf:st_bbox]{sf::bbox}}, \code{crs} as object of class \code{\link[sf:coerce-methods]{sf::crs}}, and
\code{network_buffer}, \code{dem_buffer} and \code{buildings_buffer} as numerical vector
of length one.
}
\description{
Define delineation parameters within an area of interest
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get default parameters in AOI
aoi <- define_aoi("Bucharest", "Dâmbovița")

# Get parameters in AOI with custom CRS
aoi <- define_aoi("Bucharest", "Dâmbovița",
                  crs = "EPSG:3844")  # National projected CRS


# Get parameters in AOI with non-default buffers
aoi <- define_aoi("Bucharest", "Dâmbovița",
                  network_buffer = 2000,
                  dem_buffer = 2000,
                  buildings_buffer = 150)
\dontshow{\}) # examplesIf}
}
