% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate-corridor.R
\name{define_aoi}
\alias{define_aoi}
\title{Define an area of interest (AoI) from a bounding box and a coordinate reference system (CRS).}
\usage{
define_aoi(bb, crs, buffer_dist = 0)
}
\arguments{
\item{bb}{A bounding box as a matrix with 4 elements: xmin, ymin, xmax, ymax}

\item{crs}{A coordinate reference system as an epsg code, e.g. 4326 for WGS84}

\item{buffer_dist}{A numeric value to buffer the area of interest}
}
\value{
An area of interest as a simple feature geometry
}
\description{
Define an area of interest (AoI) from a bounding box and a coordinate reference system (CRS).
}
