% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate.R
\name{delineate}
\alias{delineate}
\title{Delineate a corridor around a river}
\usage{
delineate(
  aoi,
  osm,
  dem = NULL,
  corridor_init = "valley",
  max_iterations = 10,
  capping_method = "shortest-path",
  angle_threshold = 100,
  corridor = TRUE,
  segments = FALSE,
  riverspace = FALSE
)
}
\arguments{
\item{aoi}{A list of delineation parameters for an area of interest, namely
\verb{$city_name}, \verb{$river_name}, \verb{$crs}, \verb{$network_buffer}, \verb{$dem_buffer}, and
\verb{$buildings_buffer}. For more info see \code{\link[=define_aoi]{define_aoi()}}.}

\item{osm}{A list with OpenStreetMap data sets within the \code{aoi}, as
objects of class \code{\link[sf:sfc]{sf::sfc}}}

\item{dem}{Digital elevation model (DEM) of the region (only used if
\code{corridor_init} is \code{"valley"})}

\item{corridor_init}{How to estimate the initial guess of the river corridor.
It can take the following values:
\itemize{
\item "valley": use the river valley boundary, as estimated from a Digital
Elevation Model (DEM) (for more info see \code{\link[=delineate_valley]{delineate_valley()}})
\item numeric or integer: use a buffer region of the given size (in meters)
around the river centerline
\item An \code{\link[sf:sf]{sf::sf}} or \code{\link[sf:sfc]{sf::sfc}} object: use the given input geometry
}}

\item{max_iterations}{Maximum number of iterations employed to refine the
corridor edges (see \code{\link[=corridor_edge]{corridor_edge()}}).}

\item{capping_method}{The method employed to connect the corridor edge end
points (i.e., to "cap" the corridor), as character vector of length one.
See \code{\link[=cap_corridor]{cap_corridor()}} for the available methods.}

\item{angle_threshold}{Only network edges forming angles above this threshold
(in degrees) are considered when forming segment edges. See
\code{\link[=delineate_segments]{delineate_segments()}} and \code{\link[rcoins:stroke]{rcoins::stroke()}}. Only used if \code{segments} is
TRUE.}

\item{corridor}{Whether to carry out the corridor delineation}

\item{segments}{Whether to carry out the corridor segmentation}

\item{riverspace}{Whether to carry out the riverspace delineation}
}
\value{
A list object of class \code{delineation} containing zero or more of the
following elements: "valley", "corridor", "segments", and "riverspace",
each as an \code{\link[sf:sfc]{sf::sfc_POLYGON}} or \code{\link[sf:sfc]{sf::sfc_MULTIPOLYGON}} object
(depending on the geometry of the input data). The list contains only
the geometries corresponding to the delineation steps that were carried out
(e.g., if \code{segments} is FALSE, the list will not contain a "segments"
element). In any case, the returned object also contains the base layers
and a list \code{aoi} with the parameters used for delineation.
}
\description{
Given a set of delineation parameters and input data, carry out corridor
delineation, corridor segmentation and/or riverspace delineation as indicated
by the user.
}
\details{
The returned \code{\link{delineation}} class is a list wrapping objects of class
\code{\link[sf:sfc]{sf::sfc_POLYGON}} or \code{\link[sf:sfc]{sf::sfc_MULTIPOLYGON}} which can be retrieved
through subsetting and converted to other common spatial classes in typical
\code{sf}- or \code{terra}-based workflows.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Define delineation parameters within area of interest
aoi <- define_aoi("Bucharest", "Dâmbovița")

# Get data
osm <- get_osm(aoi)
dem <- get_dem(aoi, osm)

# Delineate with defaults
delineate(aoi, osm, dem)

# Carry out all delineations
delineate(aoi, osm, dem, segments = TRUE, riverspace = TRUE)
\dontshow{\}) # examplesIf}
}
