% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate-corridor.R
\name{delineate_corridor}
\alias{delineate_corridor}
\title{Delineate a corridor around a river.}
\usage{
delineate_corridor(place, river, crs = NULL)
}
\arguments{
\item{place}{A place name as a string}

\item{river}{A river name as a string}

\item{crs}{A coordinate reference system as an epsg code, e.g. 4326 for WGS84}
}
\value{
A simple feature geometry
}
\description{
Delineate a corridor around a river.
}
