% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate.R
\name{delineate_corridor}
\alias{delineate_corridor}
\title{Delineate a corridor around a river.}
\usage{
delineate_corridor(
  city_name,
  river_name,
  crs = NULL,
  bbox_buffer = NULL,
  initial_method = "buffer",
  capping_method = "direct",
  segments = FALSE,
  riverspace = FALSE
)
}
\arguments{
\item{city_name}{A place name as a string}

\item{river_name}{A river name as a string}

\item{crs}{The projected Coordinate Reference System (CRS) to use. If not
provided, the suitable Universal Transverse Mercator (UTM) CRS is selected}

\item{bbox_buffer}{Add a buffer region to the city boundary to avoid edge
effects close to its limits}

\item{initial_method}{The method employed to define the initial river
corridor geometry. See \code{\link[=initial_corridor]{initial_corridor()}} for the available methods}

\item{capping_method}{The method employed to connect the corridor edge end
points (i.e. to "cap" the corridor). See \code{\link[=cap_corridor]{cap_corridor()}} for
the available methods}

\item{segments}{Whether to carry out the corridor segmentation}

\item{riverspace}{Whether to carry out the riverspace delineation}
}
\value{
A simple feature geometry
}
\description{
Delineate a corridor around a river.
}
