% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{delineate_corridor}
\alias{delineate_corridor}
\title{Delineate a river corridor on a spatial network.}
\usage{
delineate_corridor(
  network,
  river_centerline,
  river_surface,
  aoi,
  initial_method = "valley",
  buffer = NULL,
  dem = NULL,
  max_iterations = 10,
  capping_method = "direct"
)
}
\arguments{
\item{network}{The spatial network to be used for the delineation}

\item{river_centerline}{A simple feature geometry representing the river
centerline}

\item{river_surface}{A simple feature geometry representing the river surface}

\item{aoi}{Area of interest as sf object or bbox}

\item{initial_method}{The method employed to define the initial river
corridor geometry. See \code{\link[=initial_corridor]{initial_corridor()}} for the available methods}

\item{buffer}{Buffer region to add to the river geometry to setup the initial
corridor (only used if \code{initial_method} is \code{"buffer"})}

\item{dem}{Digital elevation model (DEM) of the region (only used if
\code{initial_method} is \code{"valley"})}

\item{max_iterations}{Maximum number of iterations employed to refine the
corridor edges (see \code{\link[=corridor_edge]{corridor_edge()}}).}

\item{capping_method}{The method employed to connect the corridor edge end
points (i.e. to "cap" the corridor). See \code{\link[=cap_corridor]{cap_corridor()}} for
the available methods}
}
\value{
A simple feature geometry representing the river corridor
}
\description{
The corridor edges on the two river banks are drawn on the provided spatial
network starting from an initial definition of the corridor (based e.g. on
the river valley).
}
\examples{
network <- rbind(bucharest_osm$streets, bucharest_osm$railways) |>
  as_network()
crs <- get_utm_zone(bucharest_osm$aoi)
aoi <- reproject(bucharest_osm$aoi, crs)
delineate_corridor(network,
                   bucharest_osm$river_centerline,
                   bucharest_osm$river_surface,
                   aoi,
                   dem = terra::unwrap(bucharest_dem))
}
