% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate.R
\name{delineate_corridor}
\alias{delineate_corridor}
\title{Delineate a corridor around a river.}
\usage{
delineate_corridor(
  city_name,
  river_name,
  crs = NULL,
  bbox_buffer = NULL,
  initial_method = "valley",
  initial_buffer = NULL,
  dem = NULL,
  max_iterations = 5,
  capping_method = "direct",
  angle_threshold = 90,
  segments = FALSE,
  riverspace = FALSE,
  ...
)
}
\arguments{
\item{city_name}{A place name as a string}

\item{river_name}{A river name as a string}

\item{crs}{The projected Coordinate Reference System (CRS) to use. If not
provided, the suitable Universal Transverse Mercator (UTM) CRS is selected}

\item{bbox_buffer}{Add a buffer region to the city boundary to avoid edge
effects close to its limits}

\item{initial_method}{The method employed to define the initial river
corridor geometry. See \code{\link[=initial_corridor]{initial_corridor()}} for the available methods}

\item{initial_buffer}{Buffer region to add to the river geometry to setup the
initial corridor (only used if \code{initial_method} is \code{"buffer"})}

\item{dem}{Digital elevation model (DEM) of the region (only used if
\code{initial_method} is \code{"valley"})}

\item{max_iterations}{Maximum number of iterations employed to refine the
corridor edges (see \code{\link[=corridor_edges]{corridor_edges()}}).}

\item{capping_method}{The method employed to connect the corridor edge end
points (i.e. to "cap" the corridor). See \code{\link[=cap_corridor]{cap_corridor()}} for
the available methods}

\item{angle_threshold}{Only network edges forming angles above this threshold
(in degrees) are considered when forming segment edges. See
\code{\link[=get_segments]{get_segments()}}  and \code{\link[rcoins:stroke]{rcoins::stroke()}}. Only used if \code{segments} is TRUE.}

\item{segments}{Whether to carry out the corridor segmentation}

\item{riverspace}{Whether to carry out the riverspace delineation}

\item{...}{Additional (optional) input arguments for retrieving the DEM
dataset (see \code{\link[=get_dem]{get_dem()}}). Only relevant if \code{initial_method} is \code{"valley"}
and \code{dem} is NULL}
}
\value{
A simple feature geometry
}
\description{
Delineate a corridor around a river.
}
