% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverspace.R
\name{delineate_riverspace}
\alias{delineate_riverspace}
\title{Delineate the space surrounding a river}
\usage{
delineate_riverspace(
  occluders,
  river,
  density = 1/50,
  ray_num = 41,
  ray_length = 100
)
}
\arguments{
\item{occluders}{Geometry of occluders}

\item{river}{List with river surface and centerline}

\item{ray_num}{Number of rays}

\item{ray_length}{Length of rays in meters}
}
\value{
Polygon geometry with the riverspace
}
\description{
Delineate the space surrounding a river
}
\examples{
\dontrun{
  delineate_riverspace(bucharest_osm$buildings, bucharest_osm$river_surface)
}
}
