% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate.R
\name{delineation}
\alias{delineation}
\title{The delineation class}
\description{
An S3 class representing the result of a delineation as returned by
\code{\link[=delineate]{delineate()}}. It extends the \code{list} class with a set of one or more of the
spatial layers \verb{$valley}, \verb{$corridor}, \verb{$segments}, and \verb{$riverspace}, as
well as the \link[=define_aoi]{area of interest} used to produce them in the
\verb{$aoi} list.
}
