% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{dem_to_cog}
\alias{dem_to_cog}
\title{Write dem to cloud optimized GeoTiff file as specified location}
\usage{
dem_to_cog(dem, fpath, output_directory = NULL)
}
\arguments{
\item{dem}{to write to file}

\item{fpath}{filepath for output. If no output directory is specified
(see below) fpath is parsed to determine
the output directory}

\item{output_directory}{where file should be written.
If specified fpath is treated as filename only.}
}
\description{
Write dem to cloud optimized GeoTiff file as specified location
}
