% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{filter_network}
\alias{filter_network}
\title{Subset a network keeping the components that intersect a target geometry.}
\usage{
filter_network(network, target, elements = "nodes")
}
\arguments{
\item{network}{A network object}

\item{target}{The target geometry}

\item{element}{The elements of the network to filter. It can be "nodes"
or "edges"}
}
\value{
A spatial network object
}
\description{
If subsetting results in multiple disconnected components, we keep the main
one.
}
\keyword{internal}
