% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{filter_network}
\alias{filter_network}
\title{Subset a network keeping the only nodes that intersect a target geometry.}
\usage{
filter_network(network, target)
}
\arguments{
\item{network}{A network object}

\item{target}{The target geometry}
}
\value{
A spatial network object
}
\description{
If subsetting results in multiple disconnected components, we keep the main
one.
}
\keyword{internal}
