% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{find_intersections}
\alias{find_intersections}
\title{Find intersections between the edges of two networks}
\usage{
find_intersections(network_1, network_2)
}
\arguments{
\item{network_1, network_2}{The two spatial network objects}
}
\value{
An object of class \code{\link[sf:sfc]{sf::sfc}}
}
\description{
Find intersections between the edges of two networks
}
\keyword{internal}
