% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{get_corridor_edges}
\alias{get_corridor_edges}
\title{Split corridor along the river to find edges on the two banks}
\usage{
get_corridor_edges(corridor, river_centerline)
}
\arguments{
\item{corridor}{The river corridor as a simple feature geometry}

\item{river_centerline}{The river centerline as a simple feature geometry}
}
\value{
Corridor edges as a simple feature geometry
}
\description{
Split corridor along the river to find edges on the two banks
}
\keyword{internal}
