% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{get_corridor_edges}
\alias{get_corridor_edges}
\title{Split corridor along the river to find edges on the two banks}
\usage{
get_corridor_edges(corridor, river)
}
\arguments{
\item{corridor}{The river corridor as a simple feature geometry}

\item{river}{The river centerline as a simple feature geometry}
}
\value{
Corridor edges as an object of class \code{\link[sf:sfc]{sf::sfc_LINESTRING}}
}
\description{
Split corridor along the river to find edges on the two banks
}
\keyword{internal}
