% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{get_dem}
\alias{get_dem}
\title{Load dem from a STAC endpoint}
\usage{
get_dem(bb, resource = "STAC", ...)
}
\arguments{
\item{bb}{A bounding box, provided either as a matrix (rows for "x", "y",
columns for "min", "max") or as a vector ("xmin", "ymin", "xmax", "ymax")}

\item{resource}{from which to source dem. Can be "STAC".
if "STAC" the parameters the following parameters
must be supplied as named parameters. If omitted defaults
will be used.}

\item{...}{Additional parameters to be passed depending on the resource.
In case \code{resource = "STAC"}, the arguments \code{endpoint} and
\code{collection} need to be passed to \code{get_stac_asset_urls()}.}
}
\value{
dem
}
\description{
Load dem from a STAC endpoint
}
