% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segments.R
\name{get_intersecting_edges}
\alias{get_intersecting_edges}
\title{Identify network edges that are intersecting a geometry.}
\usage{
get_intersecting_edges(network, geometry)
}
\arguments{
\item{network}{A spatial network object}

\item{geometry}{A simple feature geometry}
}
\value{
Indices of the edges intersecting the geometry as a vector
}
\description{
Identify network edges that are intersecting a geometry.
}
