% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_latlon}
\alias{get_latlon}
\title{Add latitude and longitude information to an sf object}
\usage{
get_latlon(x)
}
\arguments{
\item{x}{Object of class sf, sfc or sfg}
}
\value{
An sf object with latitude and longitude
}
\description{
Add latitude and longitude information to an sf object (WGS84)
}
