% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata.R
\name{get_osm_buildings}
\alias{get_osm_buildings}
\title{Get OpenStreetMap buildings}
\usage{
get_osm_buildings(bb, crs)
}
\arguments{
\item{bb}{Bounding box of class \code{bbox}}

\item{crs}{Coordinate reference system as EPSG code}
}
\value{
An sf object with the buildings
}
\description{
Get OpenStreetMap buildings
}
