% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata.R
\name{get_osm_railways}
\alias{get_osm_railways}
\title{Get OpenStreetMap railways}
\usage{
get_osm_railways(bb, crs = NULL)
}
\arguments{
\item{bb}{Bounding box of class \code{bbox}}

\item{crs}{Coordinate reference system as EPSG code}
}
\value{
An sf object with the railways
}
\description{
Get OpenStreetMap railways
}
\examples{
bb <- get_osm_bb("Bucharest")
crs <- get_utm_zone(bb)
get_osm_railways(bb, crs)
}
