% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata.R
\name{get_osmdata}
\alias{get_osmdata}
\title{Retrieve OpenStreetMap data for a given location}
\usage{
get_osmdata(
  city_name,
  river_name,
  buffer_in_m = NULL,
  crs = NULL,
  force_download = FALSE
)
}
\arguments{
\item{city_name}{A character string with the name of the city.}

\item{river_name}{A character string with the name of the river.}

\item{buffer_in_m}{An integer with the buffer size in meters
around the river center line.}

\item{crs}{An integer with the EPSG code for the projection. If no CRS is
specified, the default is the UTM zone for the city.}

\item{force_download}{Download data even if cached data is available}
}
\value{
An list with the retrieved OpenStreetMap data sets for the
given location
}
\description{
Retrieve OpenStreetMap data for a given location, including
the city boundary, the river centreline and surface, the streets, and the
railways.
}
\examples{
get_osmdata("Bucharest", "Dâmbovița", 100)
}
