% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata.R
\name{get_osmdata}
\alias{get_osmdata}
\title{Retrieve OpenStreetMap data for a given location}
\usage{
get_osmdata(bb, city_name, river_name, crs = NULL)
}
\arguments{
\item{bb}{Bounding box defining the area of interest}

\item{city_name}{A character string with the name of the city.}

\item{river_name}{A character string with the name of the river.}

\item{crs}{An integer with the EPSG code for the projection. If no CRS is
specified, the default is the UTM zone for the city.}
}
\value{
An list with the retrieved OpenStreetMap data sets for the
given location
}
\description{
Retrieve OpenStreetMap data for a given location, including
the city boundary, the river centreline and surface, the streets, and the
railways.
}
\examples{
get_osmdata("Bucharest", "Dambovita", buffer = 2000)
}
