% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retreive-osm.R
\name{get_osmdata}
\alias{get_osmdata}
\title{Retrieve OpenStreetMap data for a given location}
\usage{
get_osmdata(name, key, value)
}
\arguments{
\item{name}{A character string with the name of the place to retrieve the bounding box}

\item{key}{A character string with the key to filter the data}

\item{value}{A character string with the value to filter the data}
}
\value{
An sf object with the retrieved OpenStreetMap data for the given location
}
\description{
Retrieve OpenStreetMap data for a given location
}
\examples{
get_osmdata("Bucharest", "waterway", "river")
}
