% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{get_slope}
\alias{get_slope}
\title{Derive slope as percentage from DEM}
\usage{
get_slope(dem)
}
\arguments{
\item{dem}{raster data of dem}
}
\value{
raster of derived slope over dem extent
}
\description{
This makes use of the terrain function of the terra package
}
\keyword{internal}
