% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_utm_zone}
\alias{get_utm_zone}
\title{Get the UTM zone of a spatial object}
\usage{
get_utm_zone(x)
}
\arguments{
\item{x}{Bounding box or geometry object}
}
\value{
The EPSG of the UTM zone
}
\description{
Get the UTM zone of a spatial object
}
