% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{load_raster}
\alias{load_raster}
\title{retrieve STAC records (of a DEM) corresponding to a list of asset urls,
crop and merge with a specified bounding box to create a dem of the
specified region}
\usage{
load_raster(bb, raster_urlpaths)
}
\arguments{
\item{bb}{A bounding box (compliant with CRiSp,
i.e. as a matrix with 4 elements: xmin, ymin, xmax, ymax)}

\item{raster_urlpaths}{a list of STAC records to be retrieved}
}
\value{
A a merged dem from retrieved assets cropped to the bounding box
}
\description{
retrieve STAC records (of a DEM) corresponding to a list of asset urls,
crop and merge with a specified bounding box to create a dem of the
specified region
}
