% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_raster}
\alias{load_raster}
\title{Load raster data from one or multiple (remote) files}
\usage{
load_raster(urlpaths, bbox = NULL)
}
\arguments{
\item{urlpaths}{Path or URL to the raster file(s)}

\item{bbox}{A bounding box}
}
\value{
Raster data as a \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} object
}
\description{
If a bounding box is provided, the file(s) are cropped for the given extent.
The resulting rasters are then merged using \code{\link[terra:merge]{terra::merge}}.
}
\keyword{internal}
