% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata.R
\name{match_osm_name}
\alias{match_osm_name}
\title{Match OpenStreetMap data by name}
\usage{
match_osm_name(osm_data, name)
}
\arguments{
\item{osm_data}{An sf object with OpenStreetMap data}

\item{name}{A character string with the name to match}
}
\value{
sf object containing only rows with filtered name
}
\description{
Match OpenStreetMap data by name
}
\keyword{internal}
