% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{nearest_node}
\alias{nearest_node}
\title{Find the node in a network that is closest to a target geometry.}
\usage{
nearest_node(network, target)
}
\arguments{
\item{network}{A network object}

\item{target}{The target geometry}
}
\value{
A node in the network as a simple feature geometry
}
\description{
Find the node in a network that is closest to a target geometry.
}
