% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata.R
\name{osmdata_as_sf}
\alias{osmdata_as_sf}
\title{Retrieve OpenStreetMap data as sf object}
\usage{
osmdata_as_sf(key, value, bb)
}
\arguments{
\item{key}{A character string with the key to filter the data}

\item{value}{A character string with the value to filter the data}

\item{bb}{A bounding box, provided either as a matrix (rows for "x", "y",
columns for "min", "max") or as a vector ("xmin", "ymin", "xmax", "ymax")}
}
\value{
An sf object with the retrieved OpenStreetMap data
}
\description{
Query the Overpass API for a key:value pair within a given bounding box.
}
