% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot.delineation}
\alias{plot.delineation}
\title{Plot a delineation object}
\usage{
\method{plot}{delineation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{delineation}. This is typically the output
of the \code{delineate()} function.}

\item{...}{Additional arguments passed to the \code{plot()} function for
the \code{segments} and \code{corridor} layers.}
}
\value{
A plot visualizing the delineation object.
}
\description{
This function provides a way to quickly visulise how the layers of a
delineation object fit together. A delineation object typically includes the
base layers \code{streets}, \code{railways}, \code{river_centerline} and \code{river_surface},
and the delineations of the \code{valley}, \code{corridor}, \code{segments}, and
\code{riverspace}. Depending on the delineation object, some of the delineation
layers may not be present and thus will not be plotted.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
bd <- delineate("Bucharest", "Dâmbovița")
plot(bd)
\dontshow{\}) # examplesIf}
}
