% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrisp-package.R
\docType{package}
\name{rcrisp-package}
\alias{rcrisp}
\alias{rcrisp-package}
\title{rcrisp: Automate the Delineation of Urban River Spaces}
\description{
Provides tools to automate the morphological delineation of riverside urban areas based on a method introduced in Forgaci (2018) \doi{10.7480/abe.2018.31}. Delineation entails the identification of corridor boundaries, segmentation of the corridor, and delineation of the river space using two-dimensional spatial information from street network data and digital elevation data in a projected CRS. The resulting delineation can be used to characterise spatial phenomena that can be related to the river as a central element.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cityriverspaces.github.io/rcrisp/}
  \item \doi{10.5281/zenodo.15793526}
  \item Report bugs at \url{https://github.com/CityRiverSpaces/rcrisp/issues}
}

}
\author{
\strong{Maintainer}: Claudiu Forgaci \email{c.forgaci@tudelft.nl} (\href{https://orcid.org/0000-0003-3218-5102}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Francesco Nattino \email{f.nattino@esciencecenter.nl} (\href{https://orcid.org/0000-0003-3286-0139}{ORCID})
}

Other contributors:
\itemize{
  \item Fakhereh Alidoost \email{f.alidoost@esciencecenter.nl} (\href{https://orcid.org/0000-0001-8407-6472}{ORCID}) [contributor]
  \item Meiert Willem Grootes \email{m.grootes@esciencecenter.nl} (\href{https://orcid.org/0000-0002-5733-4795}{ORCID}) [contributor]
  \item Netherlands eScience Center \email{info@esciencecenter.nl} [funder]
}

}
\keyword{internal}
