% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reproject}
\alias{reproject}
\title{Reproject a raster or vector dataset to the specified
coordinate reference system (CRS)}
\usage{
reproject(x, crs, ...)
}
\arguments{
\item{x}{Raster or vector object}

\item{crs}{CRS to be projected to}

\item{...}{Optional arguments for raster or vector reproject functions}
}
\value{
Object reprojected to specified CRS
}
\description{
Reproject a raster or vector dataset to the specified
coordinate reference system (CRS)
}
