% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{river_buffer}
\alias{river_buffer}
\title{Draw a corridor as a fixed buffer region around a river.}
\usage{
river_buffer(river, buffer_distance, bbox = NULL, side = NULL)
}
\arguments{
\item{river}{A simple feature geometry representing the river}

\item{buffer_distance}{Size of the buffer (in the river's CRS units)}

\item{bbox}{Bounding box defining the extent of the area of interest}

\item{side}{Whether to generate a single-sided buffer with a "flat" end.
This is only applicable if \code{river} is a (multi)linestring geometry.
Choose between \code{NULL} (double-sided), \code{"right"} and \code{"left"}}
}
\value{
An object of class \code{\link[sf:sfc]{sf::sfc_POLYGON}}
}
\description{
The river geometry may consist of multiple spatial features, these are
optionally cropped using the area of interest, then merged after applying the
buffer.
}
\keyword{internal}
