% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridor.R
\name{river_buffer}
\alias{river_buffer}
\title{Draw a corridor as a fixed buffer region around a river.}
\usage{
river_buffer(river, buffer, bbox = NULL)
}
\arguments{
\item{river}{A simple feature geometry representing the river}

\item{buffer}{Size of the buffer (in the river's CRS units)}

\item{bbox}{Bounding box defining the extent of the area of interest}
}
\value{
A simple feature geometry
}
\description{
Draw a corridor as a fixed buffer region around a river.
}
