% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{shortest_path}
\alias{shortest_path}
\title{Find shortest path between a pair of nodes in the network.}
\usage{
shortest_path(network, from, to, weights = "weight")
}
\arguments{
\item{network}{A spatial network object}

\item{from}{Start node}

\item{to}{End node}

\item{weights}{Name of the column in the network edge table from where to
take the weigths}
}
\value{
A simple feature geometry
}
\description{
Find shortest path between a pair of nodes in the network.
}
\keyword{internal}
