% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate-corridor.R
\name{simplify_network}
\alias{simplify_network}
\title{Simplify a street network by removing multiple edges and loops.}
\usage{
simplify_network(net)
}
\arguments{
\item{net}{A network object}
}
\value{
A simplifed network object
}
\description{
Simplify a street network by removing multiple edges and loops.
}
