% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{simplify_network}
\alias{simplify_network}
\title{Simplify a spatial network by removing multiple edges and loops.}
\usage{
simplify_network(network)
}
\arguments{
\item{network}{A network object}
}
\value{
A simplifed network object
}
\description{
Simplify the graph, removing loops and double-edge connections following
\link[=https://luukvdmeer.github.io/sfnetworks/articles/sfn02_preprocess_clean.html#simplify-network]{this approach}.
When dropping multiple edges, keep the shortest ones.
}
