% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valley.R
\name{smooth_dem}
\alias{smooth_dem}
\title{spatially smooth dem by (window) filtering}
\usage{
smooth_dem(dem, method = "median", window = 5)
}
\arguments{
\item{dem}{raster data of dem}

\item{method}{smoothing function to be used, e.g. "median".
As accepted by focal}

\item{window}{size of smoothing kernel}
}
\value{
filtered dem
}
\description{
spatially smooth dem by (window) filtering
}
