% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delineate-corridor.R
\name{split_aoi}
\alias{split_aoi}
\title{Split the area of interest (AoI) by a river.}
\usage{
split_aoi(aoi, river)
}
\arguments{
\item{aoi}{Area of interest as a simple feature}

\item{river}{River as a simple feature}
}
\value{
A simple feature geometry set of two areas of interest
}
\description{
Split the area of interest (AoI) by a river.
}
